% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_calibration_data}
\alias{climate_calibration_data}
\title{Climate Calibration Data}
\format{
\subsection{\code{climate_calibration_data}}{

A data frame with 92 rows and 3 columns:
\describe{
\item{Site}{Locality name}
\item{MAT}{Mean Annual Temperature (celsius)}
\item{MAP}{Mean Annual Precipitation (mm)}
}
}
}
\source{
Peppe et al. 2011
}
\usage{
climate_calibration_data
}
\description{
Temperature and precipitation data associated with the modern localities used to calibrate the DiLP model
}
\references{
\itemize{
\item Peppe, D.J., Royer, D.L., Cariglino, B., Oliver, S.Y., Newman, S., Leight, E., Enikolopov, G., Fernandez-Burgos, M., Herrera, F., Adams, J.M., Correa, E., Currano, E.D., Erickson, J.M., Hinojosa, L.F., Hoganson, J.W., Iglesias, A., Jaramillo, C.A., Johnson, K.R., Jordan, G.J., Kraft, N.J.B., Lovelock, E.C., Lusk, C.H., Niinemets, Ü., Peñuelas, J., Rapson, G., Wing, S.L. and Wright, I.J. (2011), Sensitivity of leaf size and shape to climate: global patterns and paleoclimatic applications. New Phytologist, 190: 724-739. https://doi.org/10.1111/j.1469-8137.2010.03615.x
}
}
\keyword{datasets}
