% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climateCalibration}
\alias{climateCalibration}
\title{Climate Calibration Data}
\format{
\subsection{\code{climateCalibration}}{

A data frame with 92 rows and 3 columns:
\describe{
\item{Site}{Locality name}
\item{MAT}{Mean Annual Temperature (celsius)}
\item{MAP}{Mean Annual Precipitation (mm)}
}
}
}
\source{
Peppe et al. 2011
}
\usage{
climateCalibration
}
\description{
Temperature and precipitation data for modern localities used to calibrate the DiLP model
}
\keyword{datasets}
