% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilp-package.R
\docType{package}
\name{dilp-package}
\alias{dilp-package}
\title{dilp: Reconstruct Paleoclimate and Paleoecology with Leaf Physiognomy}
\description{
Use leaf physiognomic methods to reconstruct mean annual temperature (MAT), mean annual precipitation (MAP), and leaf dry mass per area (Ma), along with other useful quantitative leaf traits. Methods in this package described in Lowe et al. (in review).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjbutrim/dilp}
  \item \url{https://mjbutrim.github.io/dilp/}
  \item Report bugs at \url{https://github.com/mjbutrim/dilp/issues}
}

}
\author{
\strong{Maintainer}: Matthew J. Butrim \email{mbutrim@uwyo.edu} [copyright holder]

Authors:
\itemize{
  \item Alexander J. Lowe
  \item Andrew G. Flynn
  \item Aly Baumgartner
  \item Daniel J. Peppe
  \item Dana L. Royer
}

}
\keyword{internal}
