% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{McAbeeExample}
\alias{McAbeeExample}
\title{McAbee Example Data}
\format{
\subsection{\code{McAbeeExample}}{

A data frame with 192 rows and 18 columns:
\describe{
\item{Site}{Stratigraphic layer}
\item{Specimen Number}{Repository number for individual fossils}
\item{Morphotype}{Morphotype the specimen belongs to}
\item{Measurer comments}{Any comments about the specimen}
\item{Margin}{Whether the margin is toothed}
\item{Petiole Width}{The width of the petiole at the basalmost point of insertion (cm)}
}
}
}
\source{
Lowe et al. 2018
}
\usage{
McAbeeExample
}
\description{
Leaf physiognomic data taken from specimens collected at the McAbee
locality in Canada somewhere by Alex at some point
}
\keyword{datasets}
