% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physiognomy_calibration_data}
\alias{physiognomy_calibration_data}
\title{Physiognomy Calibration Data}
\format{
\subsection{\code{physiognomy_calibration_data}}{

A data frame with 92 rows and 12 columns:
\describe{
\item{Site}{Locality name}
\item{Leaf.area}{Average leaf area at site}
\item{FDR}{Feret diameter:Feret length. Describes leaf linearity compared to a circle}
\item{Perimeter.ratio}{Ratio - Raw blade perimeter:Internal raw blade perimeter}
\item{TC.P}{Ratio - Tooth count:Perimeter}
\item{TC.IP}{Ratio - Tooth count:Internal perimeter}
\item{Avg.TA}{Average area of a primary tooth}
\item{TA.BA}{Ratio - Tooth area:Blade area}
\item{TA.P}{Ratio - Tooth area:Perimeter}
\item{TA.IP}{Ratio - Tooth area:Internal perimeter}
\item{TC.BA}{Ratio - Tooth count:Blade area}
\item{Margin}{Percentage of untoothed species at the site}
}
}
}
\source{
Lowe et al. 2024
}
\usage{
physiognomy_calibration_data
}
\description{
Leaf physiognomic data taken from modern localities used to calibrate the DiLP model
}
\references{
\itemize{
\item Lowe, A. J., A. G. Flynn, M. J. Butrim, A. Baumgartner, D. J. Peppe, and D. L. Royer. 2024. Reconstructing terrestrial paleoclimate and paleoecology using Digital Leaf Physiognomy (DiLP) and Leaf Mass per Area. JoVE. In Review.
}
}
\keyword{datasets}
