% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physiognomyCalibration}
\alias{physiognomyCalibration}
\title{Physiognomy Calibration Data}
\format{
\subsection{\code{physiognomyCalibration}}{

A data frame with 92 rows and 12 columns:
\describe{
\item{Site}{Locality name}
\item{Leaf.area}{Average leaf area at site}
\item{FDR}{Feret diameter:Feret length. Describes leaf linearity compared to a circle}
\item{Perimeter.ratio}{Ratio - Raw blade perimeter:Internal raw blade perimeter}
\item{TC.P}{Ratio - Tooth count:Perimeter}
\item{TC.IP}{Ratio - Tooth count:Internal perimeter}
\item{Avg.TA}{Average area of a primary tooth}
\item{TA.BA}{Ratio - Tooth area:Blade area}
\item{TA.P}{Ratio - Tooth area:Perimeter}
\item{TA.IP}{Ratio - Tooth area:Internal perimeter}
\item{TC.BA}{Ratio - Tooth count:Blade area}
\item{Margin}{Percentage of untoothed species at the site}
}
}
}
\source{
Peppe et al. 2011
}
\usage{
physiognomyCalibration
}
\description{
Leaf physiognomic data taken from modern localities used to calibrate the DiLP model
}
\keyword{datasets}
