% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{royer_site_mean_ma}
\alias{royer_site_mean_ma}
\title{Site-mean (Royer) LMA parameters}
\format{
An object of class \code{list} of length 8.
}
\usage{
royer_site_mean_ma
}
\description{
Site-mean (Royer) LMA parameters
}
\keyword{datasets}
