% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{royer_species_mean_ma}
\alias{royer_species_mean_ma}
\alias{royer_site_mean_ma}
\alias{lowe_site_mean_ma}
\alias{lowe_site_variance_ma}
\title{LMA reconstruction parameters}
\format{
\describe{
\item{royer_species_mean_ma:}{
\itemize{
\item stat = "mean",
\item regression_slope = 0.382,
\item y_intercept = 3.070,
\item unexplained_mean_square = 0.032237,
\item sample_size_calibration = 667,
\item mean_log_petiole_metric_calibration = -3.011,
\item sum_of_squares_calibration = 182.1,
\item critical_value = 1.964}
}

\item{royer_site_mean_ma:}{
\itemize{
\item stat = "mean",
\item regression_slope = 0.429,
\item y_intercept = 3.214,
\item unexplained_mean_square = 0.005285,
\item sample_size_calibration = 25,
\item mean_log_petiole_metric_calibration = -2.857,
\item sum_of_squares_calibration = 5.331,
\item critical_value = 2.069}
}

\item{lowe_site_mean_ma:}{
\itemize{
\item stat = "mean",
\item regression_slope = 0.345,
\item y_intercept = 2.954,
\item unexplained_mean_square = 0.01212861,
\item sample_size_calibration = 70,
\item mean_log_petiole_metric_calibration = -2.902972,
\item sum_of_squares_calibration = 1.154691,
\item critical_value = 1.995469}
}

\item{lowe_site_variance_ma:}{
\itemize{
\item stat = "variance",
\item regression_slope = 0.302,
\item y_intercept = 5.028,
\item unexplained_mean_square = 0.1713672,
\item sample_size_calibration = 70,
\item mean_log_petiole_metric_calibration = -5.97104,
\item sum_of_squares_calibration = 5.085184,
\item critical_value = 1.995469}
}
}

An object of class \code{list} of length 8.

An object of class \code{list} of length 8.

An object of class \code{list} of length 8.
}
\usage{
royer_species_mean_ma

royer_site_mean_ma

lowe_site_mean_ma

lowe_site_variance_ma
}
\description{
Different sets of parameters used for morphospecies and site-level LMA reconstructions.
}
\references{
\itemize{
\item Royer, D. L., L. Sack, P. Wilf, C. H. Lusk, G. J. Jordan, Ulo Niinemets, I. J. Wright, et al. 2007. Fossil Leaf Economics Quantified: Calibration, Eocene Case Study, and Implications. Paleobiology 33: 574–589
\item Lowe, A. J., D. L. Royer, D. J. Wieczynski, M. J. Butrim, T. Reichgelt, L. Azevedo-Schmidt, D. J. Peppe, et al. 2024. Global patterns in community-scale leaf mass per area distributions of woody non-monocot angiosperms and their utility in the fossil record. In review.
}
}
\keyword{datasets}
