% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_regressions}
\alias{view_regressions}
\title{View preloaded regressions}
\usage{
view_regressions(type)
}
\arguments{
\item{type}{Must be either "dilp", "lma", temp", or "precip".}
}
\value{
A data frame containing the parameters for each available regression of the selected type.
}
\description{
View preloaded regressions
}
\examples{
view_regressions("dilp")
}
