% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Whittaker_biomes}
\alias{Whittaker_biomes}
\title{Whittaker Biome Data}
\format{
\subsection{\code{Whittaker_biomes}}{

A data frame with points mapping out Whittaker biome delineations.
}
}
\source{
github.com/valentinitnelav/plotbiomes
}
\usage{
Whittaker_biomes
}
\description{
Delineations of Whittaker biomes from github.com/valentinitnelav/plotbiomes
}
\references{
*Valentin Ștefan, & Sam Levin. (2018). plotbiomes: R package for plotting Whittaker biomes with ggplot2 (v1.0.0). Zenodo. https://doi.org/10.5281/zenodo.7145245
}
\keyword{datasets}
