% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lma_functions.R
\name{calc_lma}
\alias{calc_lma}
\title{Generate leaf mass per area results}
\usage{
calc_lma(data, params, resolution = "species")
}
\arguments{
\item{data}{Must include "petiole metric" or some combination of columns to
calculate petiole metric such as "Blade Area", "Petiole Area", and "Petiole Width",
or "Leaf Area" and "Petiole Width".  If calculating morphospecies-mean LMA
must include "Site" and "Morphotype" columns. If calculating species-mean LMA
must include a "Site' column.}

\item{params}{A list of regression parameters. Must contain "stat",
"regression_slope", "y_intercept", "unexplained_mean_square", "sample_size_calibration"
"mean_log_petiole_metric_calibration", "sum_of_squares_calibration", and "critical_value".

royer_species_mean_ma, royer_site_mean_ma, lowe_site_mean_ma, and
lowe_site_variance_ma are pre-loaded lists of parameters.}

\item{resolution}{Either "species" or "site".  Informs whether the function
should calculate morphospecies-mean LMA values ("species") or site-mean/site-
variance LMA values ("site").  If resolution = "site", data must already be
in the form of species-mean LMA.}
}
\value{
A table with LMA results
}
\description{
\code{calc_lma()} will typically only be called internally by \code{\link[=lma]{lma()}}. It provides
the flexibility to use custom regression parameters to calculate leaf mass
per area (LMA).
}
\examples{
#Calculate morphospecies-mean LMA values with the parameters from Royer et al. (2007)
results <-  calc_lma(McAbeeExample,
                     params = list(stat = "mean",
                                   regression_slope = 0.382,
                                   y_intercept = 3.070,
                                   unexplained_mean_square = 0.032237,
                                   sample_size_calibration = 667,
                                   mean_log_petiole_metric_calibration = -3.011,
                                   sum_of_squares_calibration = 182.1,
                                   critical_value = 1.964),
                      resolution = "species")

#Calculate site-mean LMA values with the parameters from Lowe et al. (2024)
site_results <- calc_lma(results,
                         params = list(stat = "mean",
                                       regression_slope = 0.345,
                                       y_intercept = 2.954,
                                       unexplained_mean_square = 0.01212861,
                                       sample_size_calibration = 70,
                                       mean_log_petiole_metric_calibration = -2.902972,
                                       sum_of_squares_calibration = 1.154691,
                                       critical_value = 1.995469),
                         resolution = "site")

}
