% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_calibration_data}
\alias{climate_calibration_data}
\title{Climate Calibration Data}
\format{
\subsection{\code{climate_calibration_data}}{

A data frame with 92 rows and 3 columns:
\describe{
\item{Site}{Locality name}
\item{MAT}{Mean Annual Temperature (celsius)}
\item{MAP}{Mean Annual Precipitation (mm)}
}
}
}
\source{
Lowe et al. 2024
}
\usage{
climate_calibration_data
}
\description{
Temperature and precipitation data for modern localities used to calibrate the DiLP model
}
\references{
\itemize{
\item Lowe, A. J., A. G. Flynn, M. J. Butrim, A. Baumgartner, D. J. Peppe, and D. L. Royer. 2024. Reconstructing terrestrial paleoclimate and paleoecology using Digital Leaf Physiognomy (DiLP) and Leaf Mass per Area. JoVE. In Review.
}
}
\keyword{datasets}
