% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilp-package.R
\docType{package}
\name{dilp-package}
\alias{dilp-package}
\title{dilp: Reconstruct Paleoclimate and Paleoecology With Leaf Physiognomy}
\description{
Use leaf physiognomic methods to reconstruct mean annual temperature (MAT), mean annual precipitation (MAP), and leaf dry mass per area (Ma), along with other useful quantitative leaf traits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjbutrim/dilp}
  \item Report bugs at \url{https://github.com/mjbutrim/dilp/issues}
}

}
\author{
\strong{Maintainer}: Matthew Butrim \email{mbutrim@uwyo.edu} (\href{https://orcid.org/0000-0002-7869-380X}{ORCID})

Authors:
\itemize{
  \item Alexander Lowe
  \item Andrew Flynn
  \item Aly Baumgartner
  \item Daniel Peppe
  \item Dana Royer
}

}
\keyword{internal}
