% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilp_functions.R
\name{dilp}
\alias{dilp}
\title{Generate DiLP results}
\usage{
dilp(specimen_data, params = dilp_parameters)
}
\arguments{
\item{specimen_data}{A data frame containing specimen level leaf physiognomic
data.}

\item{params}{A list of parameters used for DiLP calculation.  Defaults to
dilp_parameters, the parameters used in Peppe et al. 2011}
}
\value{
A list of tables that includes all pertinent DiLP
information:
\itemize{
\item processed_leaf_data: the full set of cleaned and newly calculated leaf
physiognomic data that is necessary for DiLP analysis.
\item processed_morphotype_data: morphospecies-site pair means for all leaf
physiognomic data
\item processed_site_data: site means for all leaf physiognomic data
\item errors: lists any specimens that may be causing common errors in DiLP
calculations
\item outliers: flags outliers in variables used for DiLP analysis that may
represent incorrect data
\item results: climate reconstructions of MAT and MAP
}
}
\description{
\code{dilp()}  processes raw leaf physiognomic data, checks for common
errors/outliers, and then returns the processed data along wiht the DiLP
paleoclimate reconstructions.
}
\examples{
results <- dilp(McAbeeExample)
}
\seealso{
\code{\link[data]{McAbeeExample}} to see how the specimen_data should be
ideally formatted.
}
