% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilp_cca.R
\name{dilp_cca}
\alias{dilp_cca}
\title{Plot site physiognomy over the physiognomic range covered by the calibration data}
\usage{
dilp_cca(
  dilp_table,
  physiognomy_calibration = physiognomyCalibration,
  climate_calibration = climateCalibration
)
}
\arguments{
\item{dilp_table}{The results of a call to \code{dilp()}}

\item{physiognomy_calibration}{A physiognomic calibration dataset. Defaults to
\code{\link[data]{physiognomyCalibration}}, the calibration dataset used by Peppe
et al. (2011)}

\item{climate_calibration}{A climate calibration dataset. Defaults to
\code{\link[data]{climateCalibration}}, the calibration dataset used by Peppe
et al. (2011)}
}
\value{
A ggplot2 plot
}
\description{
\code{dilp_cca} plots a canonical correspondence ordination of the leaf physiognomic
data used in the calibration dataset of Peppe et al. (2011).  Fossil sites
are superimposed on the cca.
}
\examples{
results <- dilp(McAbeeExample)
dilp_cca(results)

}
