% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dilp_parameters}
\alias{dilp_parameters}
\title{DiLP parameters}
\format{
\describe{
\item{dilp_parameters:}{
\itemize{
\item MAT.MLR.M = 0.21,
\item MAT.MLR.FDR = 42.296,
\item MAT.MLR.TC.IP = -2.609,
\item MAT.MLR.constant = -16.004,
\item MAT.MLR.error = 4,
\item MAT.SLR.M = 0.204,
\item MAT.SLR.constant = 4.6,
\item MAT.SLR.error = 4.9,
\item MAP.MLR.LA = 0.298,
\item MAP.MLR.TC.IP = 0.279,
\item MAP.MLR.PR = -2.717,
\item MAP.MLR.constant = 3.033,
\item MAP.MLR.SE = 0.6,
\item MAP.SLR.LA = 0.283,
\item MAP.SLR.constant = 2.92,
\item MAP.SLR.SE = 0.61}
}
}
}
\usage{
dilp_parameters
}
\description{
Parameters used in multi-linear and single linear regressions for mean annual
temperature (MAT) and mean annual precipitation (MAP) reconstructions.
}
\references{
\itemize{
\item Peppe, D.J., Royer, D.L., Cariglino, B., Oliver, S.Y., Newman, S., Leight, E., Enikolopov, G., Fernandez-Burgos, M., Herrera, F., Adams, J.M., Correa, E., Currano, E.D., Erickson, J.M., Hinojosa, L.F., Hoganson, J.W., Iglesias, A., Jaramillo, C.A., Johnson, K.R., Jordan, G.J., Kraft, N.J.B., Lovelock, E.C., Lusk, C.H., Niinemets, Ü., Peñuelas, J., Rapson, G., Wing, S.L. and Wright, I.J. (2011), Sensitivity of leaf size and shape to climate: global patterns and paleoclimatic applications. New Phytologist, 190: 724-739. https://doi.org/10.1111/j.1469-8137.2010.03615.x
}
}
\keyword{datasets}
