% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilp_whittaker.R
\name{dilp_whittaker}
\alias{dilp_whittaker}
\title{Plot climate reconstructions on a Whittaker Biome plot}
\usage{
dilp_whittaker(climate_data)
}
\arguments{
\item{climate_data}{A data frame containing either the direct output of a \code{dilp()}
call, or the $results tab from that output.
Can also be a data frame with the following required columns:
\itemize{
\item site
\item MAT.MLR
\item MAT.MLR.error
\item MAP.MLR
\item MAP.MLR.error.minus
\item MAP.MLR.error.plus
}}
}
\value{
A modifiable ggplot with dilp climate-reconstructed sites plotted
onto a Whittaker diagram.
}
\description{
\code{dilp_whittaker()} plots \code{dilp()} outputs onto a Whittaker Biome plot.  Base
Whittaker Plot from the \href{https://github.com/valentinitnelav/plotbiomes}{plotbiomes}
package by Ștefan Valentin and Sam Levin.
}
\examples{
results <- dilp(McAbeeExample)
dilp_whittaker(results)

}
\references{
Valentin Ștefan, & Sam Levin. (2018). plotbiomes: R package for plotting Whittaker biomes with ggplot2 (v1.0.0). Zenodo. https://doi.org/10.5281/zenodo.7145245
}
