% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilp_whittaker.R
\name{dilp_whittaker}
\alias{dilp_whittaker}
\title{Plot climate reconstructions on a Whittaker Biome plot}
\usage{
dilp_whittaker(climate_data)
}
\arguments{
\item{climate_data}{A data frame containing the cliamte results from a \code{dilp()} call, or
the base output of \code{dilp()}.  Can also be a data frame with the following columns:
\itemize{
\item site
\item MAT.MLR
\item MAT.MLR.error
\item MAP.MLR
\item MAP.MLR.error.minus
\item MAP.MLR.error.plus
}}
}
\value{
A ggplot with dilp climate-reconstructed sites plotted onto a Whittaker diagram.
}
\description{
Plot climate reconstructions on a Whittaker Biome plot
}
\examples{
results <- dilp(McAbeeExample)
dilp_whittaker(results)

}
