% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physiognomyCalibration}
\alias{physiognomyCalibration}
\title{Physiognomy Calibration Data}
\format{
\subsection{\code{physiognomyCalibration}}{

A data frame with 92 rows and 12 columns:
\describe{
\item{Site}{Locality name}
\item{Leaf.area}{Average leaf area at site}
\item{FDR}{Average Feret Diameter Ratio}
\item{Perimeter.ratio}{Average ratio between toothed and untoothed perimeter}
}
}
}
\source{
Peppe et al. 2011
}
\usage{
physiognomyCalibration
}
\description{
Leaf physiognomic data taken from modern localities used to calibrate the DiLP model
}
\keyword{datasets}
